Move server startup code to `main`
